# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Memories API",
        "description": "Documentação da API usando Swagger",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8005",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/pages/{hash}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Lista de páginas a partir do hash do usuário",
                "description": "Retorna todas as páginas do usuário com o hash_id",
                "operationId": "411e4531cea9dfbd33ca3c07bf886825",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash ID associado ao usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista das Páginas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Page"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/page/novo": {
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Cria uma nova página",
                "description": "Salva uma nova página do usuário",
                "operationId": "ebdecb4dc27cc332514eb6bd5986c5d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "img_01",
                                    "img_02",
                                    "img_03",
                                    "descricao"
                                ],
                                "properties": {
                                    "img_01": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "img_02": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "img_03": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição da página"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Página criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Fotos enviadas com sucesso!"
                                        },
                                        "path01": {
                                            "type": "string",
                                            "example": "https://example.com/foto1.jpg"
                                        },
                                        "path02": {
                                            "type": "string",
                                            "example": "https://example.com/foto2.jpg"
                                        },
                                        "path03": {
                                            "type": "string",
                                            "example": "https://example.com/foto3.jpg"
                                        },
                                        "page_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "hash_id": {
                                            "type": "string",
                                            "example": "4243242erwr"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/page/{id}": {
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Atualiza a página",
                "description": "Atualiza uma página com o id",
                "operationId": "d06105b72d538ac5e377908a971e2b63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da página a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "img_01",
                                    "img_02",
                                    "img_03",
                                    "descricao"
                                ],
                                "properties": {
                                    "img_01": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "img_02": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "img_03": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Nesse fizemos nossa primeira ..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Página atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Página atualizada com sucesso"
                                        },
                                        "path01": {
                                            "type": "string",
                                            "example": "https://example.com/foto1.jpg"
                                        },
                                        "path02": {
                                            "type": "string",
                                            "example": "https://example.com/foto2.jpg"
                                        },
                                        "path03": {
                                            "type": "string",
                                            "example": "https://example.com/foto3.jpg"
                                        },
                                        "page_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "hash_id": {
                                            "type": "string",
                                            "example": "42523425234"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Deleta página",
                "description": "Deleta página a partir do id",
                "operationId": "11aa7194b4958c886f5a0baf34b02659",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da página para ser deletada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Página deletada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Página deletada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Page": {
                "title": "Page",
                "description": "Modelo de uma Page",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "img_01": {
                        "type": "string",
                        "example": "https://example.com/img_01.jpg"
                    },
                    "img_02": {
                        "type": "string",
                        "example": "https://example.com/img_02.jpg"
                    },
                    "img_03": {
                        "type": "string",
                        "example": "https://example.com/img_03.jpg"
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Declaração de amor sobre o dia que casamos ..."
                    },
                    "hash_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Use o formato Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}